+++ lib/parser/models/config/app.g.dart
@@ -6,11 +6,17 @@ part of 'app.dart';
 // JsonSerializableGenerator
 // **************************************************************************
 
-App _$AppFromJson(Map json) => App(
-      android: json['android'] == null
-          ? null
-          : Android.fromJson(Map<String, dynamic>.from(json['android'] as Map)),
-      ios: json['ios'] == null
-          ? null
-          : IOS.fromJson(Map<String, dynamic>.from(json['ios'] as Map)),
-    );
+App _$AppFromJson(Map json) {
+  $checkKeys(
+    json,
+    disallowNullValues: const ['android', 'ios'],
+  );
+  return App(
+    android: json['android'] == null
+        ? null
+        : Android.fromJson(Map<String, dynamic>.from(json['android'] as Map)),
+    ios: json['ios'] == null
+        ? null
+        : IOS.fromJson(Map<String, dynamic>.from(json['ios'] as Map)),
+  );
+}