+++ lib/parser/models/flavors/flavor.g.dart
@@ -9,13 +9,16 @@ part of 'flavor.dart';
 Flavor _$FlavorFromJson(Map json) {
   $checkKeys(
     json,
-    requiredKeys: const ['app', 'android', 'ios'],
+    requiredKeys: const ['app'],
     disallowNullValues: const ['app', 'android', 'ios'],
   );
   return Flavor(
     app: App.fromJson(Map<String, dynamic>.from(json['app'] as Map)),
-    android:
-        Android.fromJson(Map<String, dynamic>.from(json['android'] as Map)),
-    ios: IOS.fromJson(Map<String, dynamic>.from(json['ios'] as Map)),
+    android: json['android'] == null
+        ? null
+        : Android.fromJson(Map<String, dynamic>.from(json['android'] as Map)),
+    ios: json['ios'] == null
+        ? null
+        : IOS.fromJson(Map<String, dynamic>.from(json['ios'] as Map)),
   );
 }