+++ lib/processors/ios/dummy_assets/ios_dummy_assets_targets_processor.dart
@@ -23,6 +23,7 @@
  * OTHER DEALINGS IN THE SOFTWARE.
  */
 
+import 'package:flutter_flavorizr/extensions/extensions_map.dart';
 import 'package:flutter_flavorizr/parser/models/flavorizr.dart';
 import 'package:flutter_flavorizr/parser/models/flavors/flavor.dart';
 import 'package:flutter_flavorizr/processors/commons/queue_processor.dart';
@@ -35,6 +36,7 @@ class IOSDummyAssetsTargetsProcessor extends QueueProcessor {
     required Flavorizr config,
   }) : super(
           config.flavors
+              .where((_, flavor) => flavor.ios != null)
               .map(
                 (String flavorName, Flavor flavor) => MapEntry(
                   flavorName,
@@ -42,7 +44,7 @@ class IOSDummyAssetsTargetsProcessor extends QueueProcessor {
                     source,
                     destination,
                     flavorName,
-                    flavor.ios,
+                    flavor.ios!,
                     config: config,
                   ),
                 ),