+++ lib/processors/ios/icons/ios_icons_processor.dart
@@ -34,12 +34,12 @@ class IOSIconsProcessor extends QueueProcessor {
   }) : super(
           config.flavors
               .where((_, flavor) =>
-                  flavor.app.icon != null || flavor.ios.icon != null)
+                  flavor.app.icon != null || flavor.ios?.icon != null)
               .map(
                 (flavorName, flavor) => MapEntry(
                   flavorName,
                   IOSIconTargetProcessor(
-                    flavor.ios.icon ?? flavor.app.icon ?? '',
+                    flavor.ios?.icon ?? flavor.app.icon ?? '',
                     flavorName,
                     config: config,
                   ),